+++ libraries/rotaryDecoder8/rotaryDecoder8.h
@@ -21,11 +21,13 @@ class rotaryDecoder8
 public:
   explicit rotaryDecoder8(const int8_t address, TwoWire *wire = &Wire);
 
-  bool     begin(uint8_t count = ROTDEC_MAX_COUNT);
+  //  note begin does not reset counters!
+  bool     begin(uint8_t deviceCount = ROTDEC_MAX_COUNT);
   bool     isConnected();
 
   uint8_t  getRECount();
-  void     reset();
+  void     reset();            //  reset all counters.
+  bool     reset(uint8_t re);  //  reset the specified counter
 
   uint16_t readInitialState();
 
@@ -35,13 +37,23 @@ public:
   bool     checkChange();
 
   //  read and update the counters
-  bool     update();         //  assumes two directions   => +1 and -1
-  bool     updateSingle();   //  assumes single direction => + ++ +++
+  bool     update();         //  assumes two directions => +1 and -1
+  bool     updateSingle();   //  assumes one direction  => +1, +2, +3
 
   //  re = rotary encoder 0..7
-  //  returns 0, false if re > 7.
+  //  returns 0 == false if re > 7.
+  //  returns the step counter of the quadrature signals (Pins A and B).
   int32_t  getValue(uint8_t re);
   bool     setValue(uint8_t re, int32_t value = 0);
+
+  //  Experimental
+  //  returns the counter of mechanical clicks (detents).
+  //          == step counter / stepPerClick.
+  int32_t  getClicks(uint8_t re);
+  bool     setClicks(uint8_t re, int32_t value);
+  //  configure per channel.
+  bool     setStepsPerClick(uint8_t re, uint8_t spc);
+  uint8_t  getStepsPerClick(uint8_t re);
 
 
   //  READ - WRITE interface
@@ -57,10 +69,11 @@ public:
 
 
 protected:
-  uint8_t   _count = 0;
+  uint8_t   _deviceCount = 0;
   uint16_t  _lastValue = 0;
   uint8_t   _lastPos[ROTDEC_MAX_COUNT] = { 0, 0, 0, 0, 0, 0, 0, 0 };
   int32_t   _encoder[ROTDEC_MAX_COUNT] = { 0, 0, 0, 0, 0, 0, 0, 0 };
+  uint8_t   _stepsPerClick[ROTDEC_MAX_COUNT] = { 1, 1, 1, 1, 1, 1, 1, 1 };
 
   uint8_t   _address;
   TwoWire * _wire;